// 2018/qiwei/src/app/g-data-vis/models/regulator.ts
export class Regulator {
    regulator_name: string;
    tap_bus: string;
    voltage_level: number;
    phase: string;
    R: number;
    X: number;
    Tap_max: number;
    Tap_min: number;
    Tap_range: number;
    PT: number;
    CT_rating: number;
    bandwidth: number;
    control_phase: string;
    startnode: string;
 
}